[QUAL] Move content of web_worker_transfer_has_filter.test.ts to the existing test file
